+++ recipes/jobs.rb
@@ -1,8 +1,11 @@
+
+
 ##############################################################
 # create job to pull down list of cookbooks
 # and
 # iterate over cookbooks and create job per cookbook in list
 ##############################################################
+
 job_name = "berkshelf"
 
 job_config = File.join(node['jenkins']['node']['home'], "#{job_name}-config.xml")